requirements.txt
aiogram==3.2.0
aio-pika==9.3.0
SQLAlchemy==2.0.23
asyncpg==0.28.0
python-dotenv==1.0.0
APScheduler==3.10.4
alembic==1.12.1
pydantic==2.4.2
pydantic-settings==2.0.3
loguru==0.7.2
requests==2.31.0
aiohttp==3.9.5 